pyproject.toml
[tool.poetry]
name = "pythonweb-hw12"
version = "0.1.0"
description = ""
authors = ["Dmytro-Tarasenko <ug.dotenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = {extras = ["email"], version = "^2.7.0"}
fastapi = "^0.110.1"
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.2"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pylint = "^3.1.0"
ruff = "^0.3.7"
faker = "^24.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
